--- style.css
@@ -1,172 +1,0 @@
-#btn-press {
-    position: relative;
-    top: 0;
-    background: #0928f2;
-    color: #fff;
-    box-shadow: 0px 6px rgb(255, 190, 186);
-    transition: 0.05s !important;
-}
-
-#btn-press:hover {
-    top: 3px;
-    box-shadow: 0px 3px rgb(146, 78, 255);
-
-}
-
-#btn-press {
-    top: 6px;
-    box-shadow: 0px 0px rgb(255, 143, 123);
-}
-
-#Title {
-    font-family: 'Segoe UI', Tahoma, Geneva, Verdana, sans-serif;
-    font-size: 45px;
-    text-shadow: 2px 2px 2px #ababab;
-    font-weight: 500;
-}
-
-#Title::selection {
-    background-color: beige;
-}
-
-#icon-card {
-    width: 64px;
-    height: 64px;
-}
-#navi{
-    
-    left: 0px;
-    height: 30px;
-    background-color: #333333;
-    color: #FFFFFF;
-    font-family: Helvetica, Verdana, Arial;
-    font-size: 16px;
-    position: fixed;
-        
-}
-.slider {
-    width: 200px;
-    height: 400px;
-    position: relative;
-    overflow: hidden;
-}
-
-.slide-ana {
-    height: 360px;
-}
-
-.slide-ana>div {
-    width: 100%;
-    height: 100%;
-    position: absolute;
-    transition: all 0.7s;
-}
-
-@media (min-width: 200px) and (max-width: 639px) {
-    .slider {
-        height: 300px;
-        width: 170px;
-    }
-}
-
-
-.card {
-    width: 220px;
-    height: 321px;
-    background: #fff;
-    border-top-right-radius: 10px;
-    overflow: hidden;
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-    position: relative;
-    box-shadow: 0 14px 26px rgba(0, 0, 0, 0.04);
-    transition: all 0.3s ease-out;
-    text-decoration: none;
-}
-
-.card:hover {
-    transform: translateY(-5px) scale(1.005) translateZ(0);
-    box-shadow: 0 24px 36px rgba(0, 0, 0, 0.11),
-        0 24px 46px var(--box-shadow-color);
-}
-
-.card:hover .overlay {
-    transform: scale(4) translateZ(0);
-}
-
-.card:hover .circle {
-    border-color: var(--bg-color-light);
-    background: var(--bg-color);
-}
-
-.card:hover .circle:after {
-    background: var(--bg-color-light);
-}
-
-.card:hover p {
-    color: var(--text-color-hover);
-}
-
-.card:active {
-    transform: scale(1) translateZ(0);
-    box-shadow: 0 15px 24px rgba(0, 0, 0, 0.11),
-        0 15px 24px var(--box-shadow-color);
-}
-
-.card p {
-    font-size: 17px;
-    color: #4C5656;
-    margin-top: 30px;
-    z-index: 1000;
-    transition: color 0.3s ease-out;
-}
-
-.circle {
-    width: 131px;
-    height: 131px;
-    border-radius: 50%;
-    background: #fff;
-    border: 2px solid var(--bg-color);
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    position: relative;
-    z-index: 1;
-    transition: all 0.3s ease-out;
-}
-
-.circle:after {
-    content: "";
-    width: 118px;
-    height: 118px;
-    display: block;
-    position: absolute;
-    background: var(--bg-color);
-    border-radius: 50%;
-    top: 7px;
-    left: 7px;
-    transition: opacity 0.3s ease-out;
-}
-
-.circle svg {
-    z-index: 10000;
-    transform: translateZ(0);
-}
-
-.overlay {
-    width: 118px;
-    position: absolute;
-    height: 118px;
-    border-radius: 50%;
-    background: var(--bg-color);
-    top: 70px;
-    left: 50px;
-    z-index: 0;
-    transition: transform 0.3s ease-out;
-}
-
-h1 {
-    font-family: system-ui, sans-serif;
-}
